feat: follow nuxt-seo docs guidelines for caching and multi-sitemaps
