+++ server/api/__sitemap__/entities.js
@@ -1,0 +1,55 @@
+import { defineCachedEventHandler } from 'nitropack/runtime';
+
+/**
+ * Dynamic sitemap URL source for entity pages.
+ * Returns deputies, parliamentary groups, and topics.
+ * Cached server-side for 1h with SWR (data updates at most daily).
+ *
+ * Recent initiatives (~2,000 newest) are handled by the sibling initiatives.js source.
+ * Full initiative inclusion is deferred to the FastAPI migration (slim endpoint).
+ */
+export default defineCachedEventHandler(
+  async (event) => {
+    const config = useRuntimeConfig(event);
+    const backendUrl = config.public.backendUrl || 'http://localhost:5000';
+    const knowledgebase = config.public.knowledgebase || 'politicas';
+
+    const params = `knowledgebase=${knowledgebase}`;
+    const urls = [];
+
+    try {
+      const [deputies, groups, topics] = await Promise.all([
+        $fetch(`${backendUrl}/deputies/?${params}`).catch(() => []),
+        $fetch(`${backendUrl}/parliamentary-groups/?${params}`).catch(() => []),
+        $fetch(`${backendUrl}/topics/?${params}`).catch(() => []),
+      ]);
+
+      for (const deputy of (deputies || [])) {
+        if (deputy.id) {
+          urls.push({ loc: `/diputados/${deputy.id}` });
+        }
+      }
+
+      for (const group of (groups || [])) {
+        if (group.id) {
+          urls.push({ loc: `/grupos/${group.id}` });
+        }
+      }
+
+      for (const topic of (topics || [])) {
+        if (topic.id) {
+          urls.push({ loc: `/tematicas/${topic.id}` });
+        }
+      }
+    } catch (err) {
+      console.warn('[sitemap:entities] Failed to fetch dynamic URLs:', err?.message);
+    }
+
+    return urls;
+  },
+  {
+    maxAge: 60 * 60, // 1 hour
+    swr: true,
+    name: 'sitemap-entities',
+  }
+);
